# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateGroupRequest:
    properties:
      avatar_url:
        example: https://storage.example.com/group-avatars/abc.jpg
        type: string
      description:
        example: Group for studying together
        maxLength: 500
        type: string
      member_emails:
        example:
        - user1@example.com
        - user2@example.com
        items:
          type: string
        maxItems: 100
        type: array
      name:
        example: Study Group
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dto.ErrorResponse:
    properties:
      metadata:
        additionalProperties:
          type: string
        type: object
      reason:
        type: string
    type: object
...
  dto.GetGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/dto.GroupDTO'
        type: array
    type: object
...
  dto.GroupDTO:
    properties:
      avatar_url:
        example: https://storage.example.com/group-avatars/abc.jpg
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        example: Group for studying together
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      member_count:
        example: 5
        type: integer
      name:
        example: Study Group
        type: string
      owner_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      pending_count:
        example: 2
        type: integer
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: Kollocol API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Get all groups the user is a member of
      parameters:
      - default: my
        description: Filter by membership type
        enum:
        - my
        - created
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetGroupsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group and invite members. avatar_url must come from
        /groups/avatar/upload.
      parameters:
      - description: Group creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GroupDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Delete a group (owner only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - groups
    get:
      description: Get details of a specific group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GroupWithMembersDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group details
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: 'Update group base info: name, description, avatar_url. Owner only.
        Setting avatar_url to empty string removes the avatar (server-side S3 cleanup).'
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GroupDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - groups
  /groups/{id}/accept:
    post:
      description: Accept a pending invitation. Marks the related group_invite notification
        as read.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GroupDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept group invitation
      tags:
      - groups
  /groups/{id}/decline:
    post:
      description: Decline a pending invitation. Marks the related group_invite notification
        as read.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Decline group invitation
      tags:
      - groups
  /groups/{id}/invite:
    post:
      consumes:
      - application/json
      description: Owner-only. Sends invitations to a list of users by email. Non-existent
        / already-invited / already-member emails are silently skipped.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Emails to invite
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InviteGroupMembersRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite members to a group
      tags:
      - groups
  /groups/{id}/kick:
    post:
      consumes:
      - application/json
      description: Owner-only. Removes a list of users by email from members and pending
        invitations. The owner cannot be kicked. Non-existent emails are silently
        skipped.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Emails to kick
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.KickGroupMembersRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kick members from a group
      tags:
      - groups
...
  /users/me/avatar/delete:
    delete:
      description: Delete current user's avatar
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete avatar
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"